43 Extract GraphQLRequestInfo.ts. Cleanup imports
