+++ src/request/GraphQLRequestInfo.ts
@@ -1,0 +1,8 @@
+import {GraphQLErrorWithStatusCode} from '..'
+
+export interface GraphQLRequestInfo {
+    query?: string
+    variables?: Readonly<Record<string, unknown>>
+    operationName?: string
+    error?: GraphQLErrorWithStatusCode
+}
